+++ aggsender/flow_base.go
@@ -1,0 +1,425 @@
+package aggsender
+
+import (
+	"context"
+	"fmt"
+	"time"
+
+	"github.com/agglayer/aggkit/agglayer"
+	"github.com/agglayer/aggkit/aggsender/db"
+	"github.com/agglayer/aggkit/aggsender/types"
+	"github.com/agglayer/aggkit/bridgesync"
+	"github.com/agglayer/aggkit/tree"
+	treeTypes "github.com/agglayer/aggkit/tree/types"
+	"github.com/ethereum/go-ethereum/common"
+	"github.com/ethereum/go-ethereum/crypto"
+)
+
+// baseFlow is a struct that holds the common logic for the different prover types
+type baseFlow struct {
+	l1InfoTreeSyncer types.L1InfoTreeSyncer
+	l2Syncer         types.L2BridgeSyncer
+	storage          db.AggSenderStorage
+
+	log types.Logger
+	cfg Config
+}
+
+// getBridgesAndClaims returns the bridges and claims consumed from the L2 fromBlock to toBlock
+func (f *baseFlow) getBridgesAndClaims(
+	ctx context.Context,
+	fromBlock, toBlock uint64,
+) ([]bridgesync.Bridge, []bridgesync.Claim, error) {
+	bridges, err := f.l2Syncer.GetBridgesPublished(ctx, fromBlock, toBlock)
+	if err != nil {
+		return nil, nil, fmt.Errorf("error getting bridges: %w", err)
+	}
+
+	if len(bridges) == 0 {
+		f.log.Infof("no bridges consumed, no need to send a certificate from block: %d to block: %d",
+			fromBlock, toBlock)
+		return nil, nil, nil
+	}
+
+	claims, err := f.l2Syncer.GetClaims(ctx, fromBlock, toBlock)
+	if err != nil {
+		return nil, nil, fmt.Errorf("error getting claims: %w", err)
+	}
+
+	return bridges, claims, nil
+}
+
+// getCertificateBuildParamsInternal returns the parameters to build a certificate
+func (f *baseFlow) getCertificateBuildParamsInternal(ctx context.Context) (*types.CertificateBuildParams, error) {
+	lastL2BlockSynced, err := f.l2Syncer.GetLastProcessedBlock(ctx)
+	if err != nil {
+		return nil, fmt.Errorf("error getting last processed block from l2: %w", err)
+	}
+
+	lastSentCertificateInfo, err := f.storage.GetLastSentCertificate()
+	if err != nil {
+		return nil, err
+	}
+
+	previousToBlock, retryCount := getLastSentBlockAndRetryCount(lastSentCertificateInfo)
+
+	if previousToBlock >= lastL2BlockSynced {
+		f.log.Infof("no new blocks to send a certificate, last certificate block: %d, last L2 block: %d",
+			previousToBlock, lastL2BlockSynced)
+		return nil, nil
+	}
+
+	fromBlock := previousToBlock + 1
+	toBlock := lastL2BlockSynced
+
+	bridges, claims, err := f.getBridgesAndClaims(ctx, fromBlock, toBlock)
+	if err != nil {
+		return nil, err
+	}
+
+	buildParams := &types.CertificateBuildParams{
+		FromBlock:           fromBlock,
+		ToBlock:             toBlock,
+		RetryCount:          retryCount,
+		LastSentCertificate: lastSentCertificateInfo,
+		Bridges:             bridges,
+		Claims:              claims,
+		CreatedAt:           uint32(time.Now().UTC().Unix()),
+	}
+
+	if buildParams.NumberOfBridges() == 0 {
+		// no bridges so no need to build the certificate
+		return nil, nil
+	}
+
+	buildParams, err = f.limitCertSize(buildParams)
+	if err != nil {
+		return nil, fmt.Errorf("error limitCertSize: %w", err)
+	}
+
+	return buildParams, nil
+}
+
+// BuildCertificate builds a certificate based on the buildParams
+// this function is the implementation of the FlowManager interface
+func (f *baseFlow) BuildCertificate(ctx context.Context,
+	buildParams *types.CertificateBuildParams) (*agglayer.Certificate, error) {
+	f.log.Infof("building certificate for %s estimatedSize=%d",
+		buildParams.String(), buildParams.EstimatedSize())
+
+	return f.buildCertificate(ctx, buildParams, buildParams.LastSentCertificate)
+}
+
+// limitCertSize limits certificate size based on the max size configuration parameter
+// size is expressed in bytes
+func (f *baseFlow) limitCertSize(fullCert *types.CertificateBuildParams) (*types.CertificateBuildParams, error) {
+	currentCert := fullCert
+	var previousCert *types.CertificateBuildParams
+	var err error
+	for {
+		if currentCert.NumberOfBridges() == 0 {
+			// We can't reduce more the certificate, so this is the minium size
+			f.log.Warnf("We reach the minium size of bridge. Certificate size: %d >max size: %d",
+				previousCert.EstimatedSize(), f.cfg.MaxCertSize)
+			return previousCert, nil
+		}
+
+		if f.cfg.MaxCertSize == 0 || currentCert.EstimatedSize() < f.cfg.MaxCertSize {
+			return currentCert, nil
+		}
+
+		// Minimum size of the certificate
+		if currentCert.NumberOfBlocks() <= 1 {
+			f.log.Warnf("reach the minium num blocks [%d to %d]. Certificate size: %d >max size: %d",
+				currentCert.FromBlock, currentCert.ToBlock, currentCert.EstimatedSize(), f.cfg.MaxCertSize)
+			return currentCert, nil
+		}
+		previousCert = currentCert
+		currentCert, err = currentCert.Range(currentCert.FromBlock, currentCert.ToBlock-1)
+		if err != nil {
+			return nil, fmt.Errorf("error reducing certificate: %w", err)
+		}
+	}
+}
+
+func (f *baseFlow) buildCertificate(ctx context.Context,
+	certParams *types.CertificateBuildParams,
+	lastSentCertificateInfo *types.CertificateInfo) (*agglayer.Certificate, error) {
+	if certParams.IsEmpty() {
+		return nil, errNoBridgesAndClaims
+	}
+
+	bridgeExits := f.getBridgeExits(certParams.Bridges)
+	importedBridgeExits, err := f.getImportedBridgeExits(ctx, certParams.Claims, certParams.L1InfoTreeRootFromWhichToProve)
+	if err != nil {
+		return nil, fmt.Errorf("error getting imported bridge exits: %w", err)
+	}
+
+	depositCount := certParams.MaxDepositCount()
+
+	exitRoot, err := f.l2Syncer.GetExitRootByIndex(ctx, depositCount)
+	if err != nil {
+		return nil, fmt.Errorf("error getting exit root by index: %d. Error: %w", depositCount, err)
+	}
+
+	height, previousLER, err := f.getNextHeightAndPreviousLER(lastSentCertificateInfo)
+	if err != nil {
+		return nil, fmt.Errorf("error getting next height and previous LER: %w", err)
+	}
+
+	meta := types.NewCertificateMetadata(
+		certParams.FromBlock,
+		uint32(certParams.ToBlock-certParams.FromBlock),
+		certParams.CreatedAt,
+	)
+
+	return &agglayer.Certificate{
+		NetworkID:           f.l2Syncer.OriginNetwork(),
+		PrevLocalExitRoot:   previousLER,
+		NewLocalExitRoot:    exitRoot.Hash,
+		BridgeExits:         bridgeExits,
+		ImportedBridgeExits: importedBridgeExits,
+		Height:              height,
+		Metadata:            meta.ToHash(),
+		AggchainProof:       certParams.AggchainProof,
+		CustomChainData:     certParams.CustomChainData,
+	}, nil
+}
+
+// createCertificateMetadata creates the metadata for the certificate
+// it returns: newMetadata + bool if the metadata is hashed or not
+func convertBridgeMetadata(metadata []byte, importedBridgeMetadataAsHash bool) ([]byte, bool) {
+	var (
+		metaData         []byte
+		isMetadataHashed bool
+	)
+
+	if importedBridgeMetadataAsHash && len(metadata) > 0 {
+		metaData = crypto.Keccak256(metadata)
+		isMetadataHashed = true
+	} else {
+		metaData = metadata
+		isMetadataHashed = false
+	}
+	return metaData, isMetadataHashed
+}
+
+// convertClaimToImportedBridgeExit converts a claim to an ImportedBridgeExit object
+func (f *baseFlow) convertClaimToImportedBridgeExit(claim bridgesync.Claim) (*agglayer.ImportedBridgeExit, error) {
+	leafType := agglayer.LeafTypeAsset
+	if claim.IsMessage {
+		leafType = agglayer.LeafTypeMessage
+	}
+	metaData, isMetadataIsHashed := convertBridgeMetadata(claim.Metadata, f.cfg.BridgeMetadataAsHash)
+
+	bridgeExit := &agglayer.BridgeExit{
+		LeafType: leafType,
+		TokenInfo: &agglayer.TokenInfo{
+			OriginNetwork:      claim.OriginNetwork,
+			OriginTokenAddress: claim.OriginAddress,
+		},
+		DestinationNetwork: claim.DestinationNetwork,
+		DestinationAddress: claim.DestinationAddress,
+		Amount:             claim.Amount,
+		IsMetadataHashed:   isMetadataIsHashed,
+		Metadata:           metaData,
+	}
+
+	mainnetFlag, rollupIndex, leafIndex, err := bridgesync.DecodeGlobalIndex(claim.GlobalIndex)
+	if err != nil {
+		return nil, fmt.Errorf("error decoding global index: %w", err)
+	}
+
+	return &agglayer.ImportedBridgeExit{
+		BridgeExit: bridgeExit,
+		GlobalIndex: &agglayer.GlobalIndex{
+			MainnetFlag: mainnetFlag,
+			RollupIndex: rollupIndex,
+			LeafIndex:   leafIndex,
+		},
+	}, nil
+}
+
+// getBridgeExits converts bridges to agglayer.BridgeExit objects
+func (f *baseFlow) getBridgeExits(bridges []bridgesync.Bridge) []*agglayer.BridgeExit {
+	bridgeExits := make([]*agglayer.BridgeExit, 0, len(bridges))
+
+	for _, bridge := range bridges {
+		metaData, isMetadataHashed := convertBridgeMetadata(bridge.Metadata, f.cfg.BridgeMetadataAsHash)
+		bridgeExits = append(bridgeExits, &agglayer.BridgeExit{
+			LeafType: agglayer.LeafType(bridge.LeafType),
+			TokenInfo: &agglayer.TokenInfo{
+				OriginNetwork:      bridge.OriginNetwork,
+				OriginTokenAddress: bridge.OriginAddress,
+			},
+			DestinationNetwork: bridge.DestinationNetwork,
+			DestinationAddress: bridge.DestinationAddress,
+			Amount:             bridge.Amount,
+			IsMetadataHashed:   isMetadataHashed,
+			Metadata:           metaData,
+		})
+	}
+
+	return bridgeExits
+}
+
+// getImportedBridgeExits converts claims to agglayer.ImportedBridgeExit objects and calculates necessary proofs
+func (f *baseFlow) getImportedBridgeExits(
+	ctx context.Context, claims []bridgesync.Claim,
+	rootFromWhichToProove *treeTypes.Root,
+) ([]*agglayer.ImportedBridgeExit, error) {
+	if len(claims) == 0 {
+		// no claims to convert
+		return []*agglayer.ImportedBridgeExit{}, nil
+	}
+
+	var (
+		importedBridgeExits = make([]*agglayer.ImportedBridgeExit, 0, len(claims))
+		rootToProve         = rootFromWhichToProove
+	)
+
+	for i, claim := range claims {
+		l1Info, err := f.l1InfoTreeSyncer.GetInfoByGlobalExitRoot(claim.GlobalExitRoot)
+		if err != nil {
+			return nil, fmt.Errorf("error getting info by global exit root: %w", err)
+		}
+
+		f.log.Debugf("claim[%d]: destAddr: %s GER: %s Block: %d Pos: %d GlobalIndex: 0x%x",
+			i, claim.DestinationAddress.String(), claim.GlobalExitRoot.String(),
+			claim.BlockNum, claim.BlockPos, claim.GlobalIndex)
+		ibe, err := f.convertClaimToImportedBridgeExit(claim)
+		if err != nil {
+			return nil, fmt.Errorf("error converting claim to imported bridge exit: %w", err)
+		}
+
+		importedBridgeExits = append(importedBridgeExits, ibe)
+
+		gerToL1Proof, err := f.l1InfoTreeSyncer.GetL1InfoTreeMerkleProofFromIndexToRoot(
+			ctx, l1Info.L1InfoTreeIndex, rootToProve.Hash,
+		)
+		if err != nil {
+			return nil, fmt.Errorf(
+				"error getting L1 Info tree merkle proof for leaf index: %d and root: %s. Error: %w",
+				l1Info.L1InfoTreeIndex, rootToProve.Hash, err,
+			)
+		}
+
+		if ibe.GlobalIndex.MainnetFlag {
+			ibe.ClaimData = &agglayer.ClaimFromMainnnet{
+				L1Leaf: &agglayer.L1InfoTreeLeaf{
+					L1InfoTreeIndex: l1Info.L1InfoTreeIndex,
+					RollupExitRoot:  claim.RollupExitRoot,
+					MainnetExitRoot: claim.MainnetExitRoot,
+					Inner: &agglayer.L1InfoTreeLeafInner{
+						GlobalExitRoot: l1Info.GlobalExitRoot,
+						Timestamp:      l1Info.Timestamp,
+						BlockHash:      l1Info.PreviousBlockHash,
+					},
+				},
+				ProofLeafMER: &agglayer.MerkleProof{
+					Root:  claim.MainnetExitRoot,
+					Proof: claim.ProofLocalExitRoot,
+				},
+				ProofGERToL1Root: &agglayer.MerkleProof{
+					Root:  rootToProve.Hash,
+					Proof: gerToL1Proof,
+				},
+			}
+		} else {
+			ibe.ClaimData = &agglayer.ClaimFromRollup{
+				L1Leaf: &agglayer.L1InfoTreeLeaf{
+					L1InfoTreeIndex: l1Info.L1InfoTreeIndex,
+					RollupExitRoot:  claim.RollupExitRoot,
+					MainnetExitRoot: claim.MainnetExitRoot,
+					Inner: &agglayer.L1InfoTreeLeafInner{
+						GlobalExitRoot: l1Info.GlobalExitRoot,
+						Timestamp:      l1Info.Timestamp,
+						BlockHash:      l1Info.PreviousBlockHash,
+					},
+				},
+				ProofLeafLER: &agglayer.MerkleProof{
+					Root: tree.CalculateRoot(ibe.BridgeExit.Hash(),
+						claim.ProofLocalExitRoot, ibe.GlobalIndex.LeafIndex),
+					Proof: claim.ProofLocalExitRoot,
+				},
+				ProofLERToRER: &agglayer.MerkleProof{
+					Root:  claim.RollupExitRoot,
+					Proof: claim.ProofRollupExitRoot,
+				},
+				ProofGERToL1Root: &agglayer.MerkleProof{
+					Root:  rootToProve.Hash,
+					Proof: gerToL1Proof,
+				},
+			}
+		}
+	}
+
+	return importedBridgeExits, nil
+}
+
+// getNextHeightAndPreviousLER returns the height and previous LER for the new certificate
+func (f *baseFlow) getNextHeightAndPreviousLER(
+	lastSentCertificateInfo *types.CertificateInfo) (uint64, common.Hash, error) {
+	if lastSentCertificateInfo == nil {
+		return 0, zeroLER, nil
+	}
+	if !lastSentCertificateInfo.Status.IsClosed() {
+		return 0, zeroLER, fmt.Errorf("last certificate %s is not closed (status: %s)",
+			lastSentCertificateInfo.ID(), lastSentCertificateInfo.Status.String())
+	}
+	if lastSentCertificateInfo.Status.IsSettled() {
+		return lastSentCertificateInfo.Height + 1, lastSentCertificateInfo.NewLocalExitRoot, nil
+	}
+
+	if lastSentCertificateInfo.Status.IsInError() {
+		// We can reuse last one of lastCert?
+		if lastSentCertificateInfo.PreviousLocalExitRoot != nil {
+			return lastSentCertificateInfo.Height, *lastSentCertificateInfo.PreviousLocalExitRoot, nil
+		}
+		// Is the first one, so we can set the zeroLER
+		if lastSentCertificateInfo.Height == 0 {
+			return 0, zeroLER, nil
+		}
+		// We get previous certificate that must be settled
+		f.log.Debugf("last certificate %s is in error, getting previous settled certificate height:%d",
+			lastSentCertificateInfo.Height-1)
+		lastSettleCert, err := f.storage.GetCertificateByHeight(lastSentCertificateInfo.Height - 1)
+		if err != nil {
+			return 0, common.Hash{}, fmt.Errorf("error getting last settled certificate: %w", err)
+		}
+		if lastSettleCert == nil {
+			return 0, common.Hash{}, fmt.Errorf("none settled certificate: %w", err)
+		}
+		if !lastSettleCert.Status.IsSettled() {
+			return 0, common.Hash{}, fmt.Errorf("last settled certificate %s is not settled (status: %s)",
+				lastSettleCert.ID(), lastSettleCert.Status.String())
+		}
+
+		return lastSentCertificateInfo.Height, lastSettleCert.NewLocalExitRoot, nil
+	}
+	return 0, zeroLER, fmt.Errorf("last certificate %s has an unknown status: %s",
+		lastSentCertificateInfo.ID(), lastSentCertificateInfo.Status.String())
+}
+
+// getLastSentBlockAndRetryCount returns the last sent block of the last sent certificate
+// if there is no previosly sent certificate, it returns 0 and 0
+func getLastSentBlockAndRetryCount(lastSentCertificateInfo *types.CertificateInfo) (uint64, int) {
+	if lastSentCertificateInfo == nil {
+		return 0, 0
+	}
+
+	retryCount := 0
+	lastSentBlock := lastSentCertificateInfo.ToBlock
+
+	if lastSentCertificateInfo.Status == agglayer.InError {
+		// if the last certificate was in error, we need to resend it
+		// from the block before the error
+		if lastSentCertificateInfo.FromBlock > 0 {
+			lastSentBlock = lastSentCertificateInfo.FromBlock - 1
+		}
+
+		retryCount = lastSentCertificateInfo.RetryCount + 1
+	}
+
+	return lastSentBlock, retryCount
+}
